### 2025-08-23 #8855 @editorialbot
[REVIEW]: Graphem - JAX: Node Influence Maximization via Geometric Embeddings
**Submitting author:** @sashakolpakov (<a href="http://orcid.org/0000-0002-6764-8894">Alexander Kolpakov</a>)
**Repository:** https://github.com/igorrivin/graphem
**Branch with paper.md** (empty if default branch): graphem-joss
**Version:** joss-0.1.0
