### 2025-09-15 igorrivin/graphem#7 @giorgospanay
[REVIEW]  Example code: minor issue converting adjacency matrix
Running the example code on README (Influence Maximization block) I get an error converting the adjacency matrix to a NetworkX object, see trace below:

```
Error testing example code (influence maximization):

Traceback (most recent call last):
  File "/Users/punksnubben/Library/Python/3.9/lib/python/site-packages/networkx/convert.py", line 144, in to_networkx_graph
    return nx.from_numpy_array(data, create_using=create_using)
  File "/Users/punksnubben/Library/Python/3.9/lib/python/site-packages/networkx/utils/backends.py", line 412, in __call__
    return self.orig_func(*args, **kwargs)
  File "/Users/punksnubben/Library/Python/3.9/lib/python/site-packages/networkx/convert_matrix.py", line 1118, in from_numpy_array
    raise nx.NetworkXError(f"Adjacency matrix not square: nx,ny={A.shape}")
networkx.exception.NetworkXError: Adjacency matrix not square: nx,ny=(1204, 2)

Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/test-graphem.py", line 25, in <module>
    G = nx.Graph(edges)
  File "/Users/punksnubben/Library/Python/3.9/lib/python/site-packages/networkx/classes/graph.py", line 370, in __init__
    convert.to_networkx_graph(incoming_graph_data, create_using=self)
  File "/Users/punksnubben/Library/Python/3.9/lib/python/site-packages/networkx/convert.py", line 146, in to_networkx_graph
    raise nx.NetworkXError(
networkx.exception.NetworkXError: Failed to interpret array as an adjacency matrix.
```

openjournals/joss-reviews#8855
