### 2025-09-15 igorrivin/graphem#7 @sashakolpakov
Addressed in https://github.com/igorrivin/graphem/pull/6

Namely, instead of
```
G = nx.Graph(edges)
```
in README.md, one has to use
```
G = nx.from_edgelist(edges)
```
since `edges` is a list of edges, not an adjacency matrix.
