### 2025-09-15 #8855 @sashakolpakov
@danielskatz Thanks, @giorgospanay ! That's a nice catch! The issue is confirmed, and now addressed  in https://github.com/igorrivin/graphem/pull/6

Namely, instead of
```
G = nx.Graph(edges)
```
in README.md, one has to use
```
G = nx.from_edgelist(edges)
```
since `edges` is a list of edges, not an adjacency matrix.
