### 2025-09-24 #8855 @sashakolpakov
@danielskatz Thanks for proofreading! I will check it out in the next 24 hours.

Other editorial questions are answered below:

>> ... Can you explain why?
Easier for me to handle the submission. @igorrivin will merge the final pull requests after the review is finished.

>> Should we be pointing to https://github.com/igorrivin/graphem as the software's home?
Yes, the root repo will be more appropriate in the long run.
