### 2025-09-25 #8855 @danielskatz
@sashakolpakov - is the paper in the current repo? We still need it, to build the PDF and do the acceptance
