### 2025-09-25 #8855 @sashakolpakov
@danielskatz The paper is in the "graphem-joss" branch.
