### 2025-09-25 #8855 @sashakolpakov
@danielskatz Thanks! Done!

Release: https://github.com/igorrivin/graphem/tree/joss-0.1.0
Zenodo: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17203428.svg)](https://doi.org/10.5281/zenodo.17203428)
