### 2025-09-25 #8855 @danielskatz
Congratulations to @sashakolpakov (Alexander Kolpakov) and co-author on your JOSS publication!!

And thanks to @leotrs & @giorgospanay for reviewing!
JOSS completely depends on volunteers, and we couldn't do this without you
